+++ src/viriatum/test/simple_test.c
@@ -1487,6 +1487,7 @@ static struct test_entry_t _simple_entries[] = {
     V_TEST_T(test_mode_options_service, "service"),
     V_TEST_T(test_bind_options_service, "service"),
     V_TEST_T(test_static_options_service, "service"),
+    V_TEST_T(test_dev_options_service, "service"),
     V_TEST_T(test_ephemeral_service, "service"),
     V_TEST_T(test_polling_service, "service"),
     V_TEST_T(test_polling_connection, "polling"),